chore: split out logic in new useImageBlobUrl hook
