+++ app/_components/DropdownMenu/index.tsx
@@ -1,4 +1,4 @@
-import { Menu, MenuButton } from '@szhsin/react-menu'
+import { Menu, MenuButton, MenuDirection } from '@szhsin/react-menu'
 import '@szhsin/react-menu/dist/index.css'
 import '@szhsin/react-menu/dist/transitions/slide.css'
 
@@ -9,7 +9,7 @@ export default function DropdownMenu({
 }: {
   children: React.ReactNode
   menuButton: React.ReactNode
-  direction?: string
+  direction?: MenuDirection
 }) {
   return (
     <Menu
