+++ app/_components/ImageView/ImageViewImage.tsx
@@ -1,43 +1,15 @@
 /* eslint-disable @next/next/no-img-element */
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { useImageView } from './ImageViewProvider'
 import CarouselImage from '../Carousel/CarouselImage'
 import Carousel from '../Carousel'
 
-const defaultImage =
-  'data:image/gif;base64,R0lGODdhAQABAJEAAAAAAB8fH////wAAACH5BAkAAAMALAAAAAABAAEAAAICTAEAOw=='
-
 // Prevents re-rendering of the same image multiple times as parent is updated
 const ImageViewImage = React.memo(() => {
-  const { artbot_id, imageBlob, imageData, setCurrentImageId } = useImageView()
+  const { imageBlob, imageData, setCurrentImageId } = useImageView()
   const { imageFiles } = imageData
 
-  const [imageUrl, setImageUrl] = useState(defaultImage)
-
-  useEffect(() => {
-    let isMounted = true // To prevent state update on unmounted component
-
-    const fetchImage = async () => {
-      if (imageBlob && isMounted) {
-        const url = URL.createObjectURL(imageBlob)
-        setImageUrl(url)
-      }
-    }
-
-    fetchImage()
-
-    return () => {
-      isMounted = false // Clean up to prevent memory leaks
-
-      if (imageUrl !== defaultImage) {
-        URL.revokeObjectURL(imageUrl) // Clean up blob URL
-      }
-    }
-
-    // imageUrl in this dep array causes issue with URL.createObjectURL re-rendering thousands of times.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [artbot_id, imageBlob])
-
+  if (!imageFiles) return null
   if (!imageBlob) return null
 
   return (
